Canary pods need to be created via Jobs
For the pia canary to work well, the pod creation needs to be on the same apiserver instance that the controller manager uses. I really hope that controller manager requests are not load balanced across all 3.
We can use a Job object so that the controller manager makes the pods.