Add support for operator driven cluster provisioning
This commit adds support for operator driven cluster provisioning. This feature is by default
disabled. It needs to be enabled via spec.clusterProvisioning.enabled: true. If this feature
is enabled, the operator will provision the cluster by calling the provision_cluster gRPC endpoint
of restate-0 with the default values (provisioning the cluster with the values it was configured
with). To avoid reprovisioning a cluster, we then update status.provisioned to true. This will
prevent future reconciliation loops from reprovisioning the cluster.

If the operator driven cluster provisioning is enabled, then we require that Restate server's
auto-provisioning is disabled. Hence, we fail if auto-provision = false is not found in the configuration
as this can lead to split brain situations.

For existing clusters where users update the CR to enable the operator driven cluster provisioning, the
operator will run the grpc provision call again. If it goes against a provisioned cluster, nothing will
happen. However, there is small risk that this call might reprovision restate-0 if it lost its disk or
wasn't provisioned before. Therefore, users need to be careful with this feature when enabling it for an
existing RestateCluster.

As part of this change, we also bumped the Rust edition to 2024 be able
to use let chaining.

This fixes #55.